requirements.txt
streamlit>=1.20.0
streamlit-folium>=0.11.1
osmnx>=1.3.0
networkx>=3.0
folium>=0.14.0
pandas>=2.0.0
geopandas>=0.14.0
numpy>=1.24.0
pydeck>=0.8.0
scikit-learn>=1.3.0
Pillow>=9.0.0
